requirements.txt
google-genai>=1.0.0
chromadb>=0.5.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
tenacity
watchdog
pdf2image
sqlalchemy>=2.0.0
python-dotenv
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
